Adds support for the SystemMediaTransportControls.
 Some playback buttons and song information are added to the windows
 volume control OSD.

 There are other features by the look of it, track position, shuffle
 etc. I'm not sure how to enable them properly so it's commented out for
 now.

+++ release.sh
@@ -49,6 +49,8 @@ log4net.dll
 ManagedBass.dll
 ManagedBass.Gapless.dll
 ManagedBass.Mix.dll
+msvcp100.dll
+msvcr100.dll
 Noesis.Javascript.dll
 System.Data.SQLite.dll
 System.IO.dll
@@ -64,6 +66,7 @@ Microsoft.Windows.Shell.dll
 
 WINDOWS="
 FoxTunes.Core.Windows.dll
+FoxTunes.Core.Windows.UWP.dll
 "
 
 ASIO="
@@ -200,6 +203,11 @@ do
 
 	for file in $WINDOWS
 	do
+		if [ ! -f "./distribution/$target/$file" ]
+		then
+			echo "SKIPPING $file" 
+			continue
+		fi
 		echo $file
 		cp "./distribution/$target/$file" "./release/$target/Plugins/windows"
 	done
